refactor hashtable
